Remove driver files entirely
